requirements.txt

python-telegram-bot
openai
requests
pytesseract
Pillow
python-dotenv
selenium
anticaptchaofficial
pdfplumber
beautifulsoup4
